# Refactor custom FPx cast