Closes #354

TODO:
- ~Check torch.compile~
- ~Benchmark before and after~

```
python torchao/prototype/mx_formats/benchmarks/bench_qdq.py
```

88410947 (main)

| elem_dtype          | use_fp4_custom_triton_dequant_kernel   |   q_time_us |   q_mem_bw_tb_s |   dq_time_us |   dq_mem_bw_tb_s |
|---------------------|----------------------------------------|-------------|-----------------|--------------|------------------|
| torch.float8_e4m3fn | False                                  |      532.20 |            0.26 |       554.40 |             0.25 |
| torch.float8_e5m2   | False                                  |      532.83 |            0.26 |       551.01 |             0.25 |
| fp6_e2m3            | False                                  |      574.66 |            0.24 |       258.41 |             0.53 |
| fp6_e3m2            | False                                  |      577.37 |            0.24 |       258.86 |             0.53 |
| fp4_e2m1            | False                                  |      682.13 |            0.17 |       254.72 |             0.45 |
| fp4_e2m1            | True                                   |    12251.34 |            0.01 |       190.39 |             0.60 |

2690b927 (this PR)

| elem_dtype          | use_fp4_custom_triton_dequant_kernel   |   q_time_us |   q_mem_bw_tb_s |   dq_time_us |   dq_mem_bw_tb_s |
|---------------------|----------------------------------------|-------------|-----------------|--------------|------------------|
| torch.float8_e4m3fn | False                                  |      531.62 |            0.26 |       552.91 |             0.25 |
| torch.float8_e5m2   | False                                  |      530.52 |            0.26 |       550.33 |             0.25 |
| fp6_e2m3            | False                                  |      572.89 |            0.24 |       551.21 |             0.25 |
| fp6_e3m2            | False                                  |      576.62 |            0.24 |       551.92 |             0.25 |
| fp4_e2m1            | False                                  |      680.27 |            0.17 |       255.10 |             0.45 |
| fp4_e2m1            | True                                   |    12248.68 |            0.01 |       191.09 |             0.60 |

Dequant is 2x slower because I replaced LUT-based denormal handling with a more generic logic. @vkuzo Should I add back the LUT-based logic (check specifically for E2M3 E3M2 E2M1)? If we are interested in performance then perhaps we can generate a LUT for all bit patterns and cache it.

UPDATE

95f4582 (this PR v2)

| elem_dtype          | use_fp4_custom_triton_dequant_kernel   |   q_time_us |   q_mem_bw_tb_s |   dq_time_us |   dq_mem_bw_tb_s |
|---------------------|----------------------------------------|-------------|-----------------|--------------|------------------|
| torch.float8_e4m3fn | False                                  |      531.54 |            0.26 |       554.03 |             0.25 |
| torch.float8_e5m2   | False                                  |      532.41 |            0.26 |       551.63 |             0.25 |
| fp6_e2m3            | False                                  |      574.66 |            0.24 |       258.28 |             0.53 |
| fp6_e3m2            | False                                  |      576.53 |            0.24 |       258.76 |             0.53 |
| fp4_e2m1            | False                                  |      682.26 |            0.17 |       517.65 |             0.22 |
| fp4_e2m1            | True                                   |    12247.99 |            0.01 |       190.48 |             0.60 |

Now FP4_E2M1 is slower lol. Feel like this should be bandwidth-limited. It might be register-limited also? Will do some profiling + make sure torch.compile run optimally. Interesting that native PyTorch float8 dequant is slower.

UPDATE 2

dcd5a05a (this PR v3)

| elem_dtype          | use_fp4_custom_triton_dequant_kernel   |   q_time_us |   q_mem_bw_tb_s |   dq_time_us |   dq_mem_bw_tb_s |
|---------------------|----------------------------------------|-------------|-----------------|--------------|------------------|
| torch.float8_e4m3fn | False                                  |      532.65 |            0.26 |       554.13 |             0.25 |
| torch.float8_e5m2   | False                                  |      532.07 |            0.26 |       551.30 |             0.25 |
| fp6_e2m3            | False                                  |      574.76 |            0.24 |       258.48 |             0.53 |
| fp6_e3m2            | False                                  |      576.85 |            0.24 |       258.36 |             0.53 |
| fp4_e2m1            | False                                  |      681.22 |            0.17 |       254.14 |             0.45 |
| fp4_e2m1            | True                                   |    12249.79 |            0.01 |       190.49 |             0.60 |

Speed recovered 😊