@vkuzo 2024-06-15
> Dequant is 2x slower because I replaced LUT-based denormal handling with a more generic logic.

2x is a sizeable regression, how about keeping the LUT for the formats we already have it for and having a generic fallback for the other formats?  People can then optimize format by format individually if they want.